refactor(gateway): extract cron application service
