+++ docs/gateway-architecture-refactor-plan.md
@@ -259,7 +259,7 @@ Domain & Infra Adapters
 - [x] PR-1 阶段 0：测试护栏
 - [x] PR-2 阶段 1：Transport 拆分
 - [x] PR-3 阶段 2A：Agent Service
-- [ ] PR-4 阶段 2B：Cron Service
+- [x] PR-4 阶段 2B：Cron Service
 - [ ] PR-5 阶段 2C：Workspace/Model Service
 - [ ] PR-6 阶段 3：Ports + Adapters
 - [ ] PR-7 阶段 4/5：SystemPrompt 收敛 + 治理
